refactor: cmd dir layout
