+++ pkg/http/healthcheck.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package main
+package http
 
 import (
 	"net/http"
