+++ pkg/http/routes.go
@@ -14,15 +14,21 @@
  * limitations under the License.
  */
 
-package main
+package http
 
 import (
+	"net/http"
+
 	"github.com/gorilla/mux"
-	"github.com/prometheus/client_golang/prometheus/promhttp"
 )
 
-// registerMetricsRouter - add handler functions for metrics.
-func registerMetricsRouter(router *mux.Router) {
-	// metrics router
-	router.Path("/metrics").Handler(promhttp.Handler())
+func RegisterRoutes() (http.Handler, error) {
+	router := mux.NewRouter().SkipClean(true).UseEncodedPath()
+	// Add healthcheck router
+	registerHealthCheckRouter(router)
+
+	// Add server metrics router
+	registerMetricsRouter(router)
+
+	return router, nil
 }
